multi-clients can  join channel but missing functionality
need to find out what data needs to be sent back from the server to the client for each Command. Right its its setup so the latest connection to the channel can see the list of all who connected to the channel before it but not after. I'm thinking its because you have to send back to each client data which updates the clients connected list. Need to finish PRVMSG stuff on reading part to finish basic functionality for READ Commands.

Can get started on WRITE commands to send stuff back to clients.

Bot connection is a bit iffy. Right now the bot can connect to the server. However on connection it only runs the Username command and does not run Nickname or JoinChannel stuff.